refactored entities User and Transactions and Category.

+++ Backend/Trifolium/src/main/java/com/alorisse/trifolium/model/entity/Category.java
@@ -12,7 +12,7 @@ public class Category {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false)
-    private Integer id;
+    private Long id;
 
     @Column(name = "title", nullable = false, length = 100)
     private String title;
@@ -23,6 +23,4 @@ public class Category {
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id")
     private User user;
-
-
 }
